+++ lib/src/home/view/screens/android/modules_android.dart
@@ -6,12 +6,11 @@ import 'package:playhouse/src/model.dart';
 
 import 'package:playhouse/src/view.dart';
 
-class OrganizationsAndroid
-    extends ScrapbookListScreen<OrganizationsList, OrganizationsFields>
+class ModulesListAndroid extends ScrapbookListScreen<ModulesList, ModuleFields>
     with ScrapbookFields {
 
   @override
-  String title = 'Organizations List';
+  String title = 'Modules List';
 
   @override
   bool useModuleType = false;
@@ -36,5 +35,5 @@ class OrganizationsAndroid
 
   @override
   List<Map<String, FieldWidgets<PlayHouseFields>>>
-      fetchData() => OrganizationsFields().field.values.toList();
+  fetchData() => ModuleFields().field.values.toList();
 }