+++ lib/src/home/view/screens/android/tasks_android.dart
@@ -6,12 +6,12 @@ import 'package:playhouse/src/model.dart';
 
 import 'package:playhouse/src/view.dart';
 
-class OrganizationsAndroid
-    extends ScrapbookListScreen<OrganizationsList, OrganizationsFields>
+class TasksListAndroid extends ScrapbookListScreen<TasksList, TasksFields>
     with ScrapbookFields {
+//
 
   @override
-  String title = 'Organizations List';
+  String title = 'Tasks List';
 
   @override
   bool useModuleType = false;
@@ -36,5 +36,5 @@ class OrganizationsAndroid
 
   @override
   List<Map<String, FieldWidgets<PlayHouseFields>>>
-      fetchData() => OrganizationsFields().field.values.toList();
+  fetchData() => TasksFields().field.values.toList();
 }