+++ lib/src/view.dart
@@ -15,7 +15,7 @@ export 'package:playhouse/src/app/view/menu/app_menu.dart';
 export 'package:playhouse/src/app/view/signin_screen.dart';
 
 /// Scrapbook main screen
-export 'package:playhouse/src/home/view/scrapbook_main.dart'; //  hide ScrapBookApp
+export 'package:playhouse/src/home/view/scrapbook_main.dart';
 
 /// AppBars
 export 'package:playhouse/src/home/view/appbar/design_build_appbar.dart';
@@ -55,11 +55,39 @@ export 'package:playhouse/src/home/view/taskscards.dart';
 export 'package:playhouse/src/home/view/page_circle_indicator.dart';
 
 /// Screens
+export 'package:playhouse/src/home/view/screens/scrapbook_screens.dart';
+export 'package:playhouse/src/home/view/screens/scrapbook_screens_widgets.dart';
+
+/// Modules
+export 'package:playhouse/src/home/view/screens/modules_list.dart';
+export 'package:playhouse/src/home/view/screens/android/modules_android.dart';
+export 'package:playhouse/src/home/view/screens/ios/modules_ios.dart';
+
+/// Submodules
+export 'package:playhouse/src/home/view/screens/submodules_list.dart';
+export 'package:playhouse/src/home/view/screens/android/submodules_android.dart';
+export 'package:playhouse/src/home/view/screens/ios/submodules_ios.dart';
+
+/// Tasks
+export 'package:playhouse/src/home/view/screens/tasks_list.dart';
+export 'package:playhouse/src/home/view/screens/android/tasks_android.dart';
+export 'package:playhouse/src/home/view/screens/ios/tasks_ios.dart';
+
+/// Organizations
 export 'package:playhouse/src/home/view/screens/organizations_list.dart';
-
 export 'package:playhouse/src/home/view/screens/android/organizations_android.dart';
-
 export 'package:playhouse/src/home/view/screens/ios/organizations_ios.dart';
+
+/// Users
+export 'package:playhouse/src/home/view/screens/users_list.dart';
+export 'package:playhouse/src/home/view/screens/android/users_android.dart';
+export 'package:playhouse/src/home/view/screens/ios/users_ios.dart';
+
+export 'package:playhouse/src/home/view/screens/users_list.dart';
+
+export 'package:playhouse/src/home/view/screens/users_list.dart';
+
+export 'package:playhouse/src/home/view/screens/users_list.dart';
 
 export 'package:playhouse/src/home/view/screens/users_list.dart';